requirements.txt
django-imagekit==4.0.2
Django==2.2.1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
